1.新增.env文件 2.优化权限更新 3.增加数据库迁移文件

+++ apps/server/src/modules/sys/sys-menu/sys-menu.service.ts
@@ -5,6 +5,7 @@ import {
   UpdateSysMenuDto,
 } from './dto/req-sys-menu.dto';
 
+import { ApiException } from '@/common/exceptions/api.exception';
 import { buildMenuTree } from '@/utils/util';
 import { Prisma } from '@prisma/client';
 import { PrismaService } from 'nestjs-prisma';
@@ -13,8 +14,14 @@ import { PrismaService } from 'nestjs-prisma';
 export class SysMenuService {
   constructor(private readonly prisma: PrismaService) {}
   async create(createSysMenuDto: CreateSysMenuDto) {
-    const { meta, menuBtns, parentId, ...other } = createSysMenuDto;
-    return await this.prisma.sysMenu.create({
+    const {
+      meta,
+      menuBtns = [],
+      parameters = [],
+      parentId,
+      ...other
+    } = createSysMenuDto;
+    return this.prisma.sysMenu.create({
       data: {
         ...other,
         parentId: parentId == 0 ? null : parentId,
@@ -24,6 +31,11 @@ export class SysMenuService {
         menuBtns: {
           createMany: {
             data: menuBtns,
+          },
+        },
+        parameters: {
+          createMany: {
+            data: parameters,
           },
         },
       },
@@ -83,6 +95,7 @@ export class SysMenuService {
       include: {
         meta: true,
         menuBtns: true,
+        parameters: true,
         parent: true,
       },
     });
@@ -96,24 +109,71 @@ export class SysMenuService {
   }
 
   async update(id: number, updateSysMenuDto: UpdateSysMenuDto) {
-    const { meta, menuBtns = [], parentId, ...other } = updateSysMenuDto;
-    const metaInfo = meta ? { update: meta } : undefined;
-    // TODO: 这里还需要删除角色菜单权限表的关系
-    return this.prisma.sysMenu.update({
-      where: {
-        id,
-      },
-      data: {
-        ...other,
-        parentId: parentId == 0 ? null : parentId,
-        meta: metaInfo,
-        menuBtns: {
-          deleteMany: {},
-          createMany: {
-            data: menuBtns,
+    const {
+      meta,
+      menuBtns = [],
+      parameters = [],
+      parentId,
+      ...other
+    } = updateSysMenuDto;
+
+    const exists = await this.prisma.sysMenu.findUnique({ where: { id } });
+    if (!exists) throw new ApiException('当前菜单信息不存在');
+    // 因为这里有关联表的缘故，所以不能全部删除再新增，需要判断数据是删除还是新增还有编辑
+    // 找出列表里带 ID 的（说明是老数据，需要 Update）
+    const updates = menuBtns.filter((btn) => btn.id != undefined);
+    // 找出列表里不带 ID 的（说明是新数据，需要 Create）
+    const creates = menuBtns.filter((btn) => !btn.id);
+
+    // 提取出“幸存者”的 ID 列表
+    const keepIds: number[] = updates.map((btn) => btn.id as number);
+    return this.prisma.$transaction(async (tx) => {
+      // 数据库里属于这个菜单，但不在 keepIds 列表里的，统统删掉
+      await tx.sysMenuBtn.deleteMany({
+        where: {
+          sysMenuId: id,
+          id: {
+            notIn: keepIds,
           },
         },
-      },
+      });
+      //   更新数据
+      if (updates.length > 0) {
+        await Promise.all(
+          updates.map((btn) => {
+            const { id, ...data } = btn;
+            return tx.sysMenuBtn.update({
+              where: { id },
+              data: data, // 只更新 name, auth
+            });
+          }),
+        );
+      }
+      if (creates.length > 0) {
+        await tx.sysMenuBtn.createMany({
+          data: creates.map((btn) => ({
+            ...btn,
+            sysMenuId: id, // 绑定父菜单
+          })),
+        });
+      }
+      const metaInfo = meta ? { update: meta } : undefined;
+      return tx.sysMenu.update({
+        where: {
+          id,
+        },
+        data: {
+          ...other,
+          parentId: parentId == 0 ? null : parentId,
+          meta: metaInfo,
+          parameters: {
+            deleteMany: {},
+            createMany: {
+              data: parameters,
+            },
+          },
+        },
+      });
     });
   }
